+++ plugin/bn128.md
@@ -1,0 +1,29 @@
+```k
+module CRYPTOGRAPHY-BN128-ELLIPTIC-CURVE
+    imports STRING-SYNTAX
+    imports INT-SYNTAX
+    imports LIST
+```
+
+The BN128 elliptic curve is defined over 2-dimensional points over the fields of zero- and first-degree polynomials modulo a large prime. (x, y) is a point on G1, whereas (x1 x x2, y1 x y2) is a point on G2, in which x1 and y1 are zero-degree coefficients and x2 and y2 are first-degree coefficients. In each case, (0, 0) is used to represent the point at infinity.
+
+-   `BN128Add` adds two points in G1 together,
+-   `BN128Mul` multiplies a point in G1 by a scalar.
+-   `BN128AtePairing` accepts a list of points in G1 and a list of points in G2 and returns whether the sum of the product of the discrete logarithm of the G1 points multiplied by the discrete logarithm of the G2 points is equal to zero.
+-   `isValidPoint` takes a point in either G1 or G2 and validates that it actually falls on the respective elliptic curve.
+
+```k
+    syntax G1Point ::= "(" Int "," Int ")"
+    syntax G2Point ::= "(" Int "x" Int "," Int "x" Int ")"
+    syntax G1Point ::= BN128Add(G1Point, G1Point) [function, hook(BN128.bn128add)]
+                     | BN128Mul(G1Point, Int)     [function, hook(BN128.bn128mul)]
+ // -------------------------------------------------------------------------------
+
+    syntax Bool ::= BN128AtePairing(List, List) [function, hook(BN128.bn128ate)]
+ // -----------------------------------------------------------------------------
+
+    syntax Bool ::= isValidPoint(G1Point) [function, hook(BN128.bn128valid)]
+                  | isValidPoint(G2Point) [function, klabel(isValidG2Point), hook(BN128.bn128g2valid)]
+ // ---------------------------------------------------------------------------------------------------
+endmodule
+```
